requirements.txt
# Core dependencies
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1


# Background tasks
celery==5.4.0
redis==5.2.1

# Scientific computing
numpy==2.0.2
psutil==5.9.8

# Development tools
django-debug-toolbar==4.3.0
django-extensions==3.2.3
ipython==8.18.1
black==24.1.1

# API Documentation
drf-yasg==1.21.7
coreapi==2.3.3
drf-spectacular==0.28.0

# JWT Authentication
djangorestframework-simplejwt==5.5.0
# Celery dependencies
amqp==5.3.1
async-timeout==5.0.1
billiard==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
kombu==5.4.2
prompt_toolkit==3.0.50
vine==5.1.0
wcwidth==0.2.13

# Other dependencies
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
channels-redis==4.1.0
channels==4.0.0
uvicorn==0.34.0
daphne==4.1.2
websockets==15.0.1
setuptools==69.1.0
